backend/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
Pillow==10.0.0
djoser==2.2.0
psycopg2-binary==2.9.7